Fixed the screen resolution and fixed the expand powerup
Confusing a bit, as Random.range() is inclusive for floats and exclusive for ints.
Also, the paddles were sometimes off screen, especially in a build.
Also, added ESC to end the game
And find it plays better when the expand and shrink powerups stay around until the next point is won.